==> Analysis/4-DataExploration.ipynb <==
# This notebook was used to conduct preliminary exploration of data
# As such it is not a part of the main workflow, but used just for additional insights

# Import the required libraries
import pandas as pd
import time
from datetime import timedelta, datetime
from sqlalchemy import create_engine
from sklearn.feature_extraction.text import CountVectorizer
import os

new_working_directory = "/Users/fatimaq/Documents/Qualitative_reddit_analysis/Output"
os.chdir(new_working_directory)



# Create SQL engine; choose the correct path where reddit.db database file is located
engine = create_engine(
    f"sqlite:////Users/fatimaq/Documents/Qualitative_reddit_analysis/Output/testredditpost.sqlite")


# Select all the data from Table A in the database
posts = pd.read_sql("""select * from my_table order by createdAt""",engine)
posts.head()

# read the total length of posts
len(posts)

posts.tail()

# Convert the Unix epoch time to human readable dates and set index
# Convert 'createdAt' column to datetime objects
posts['createdAt'] = pd.to_datetime(posts['createdAt'])

# Convert datetime objects to Unix epoch timestamps
posts['createdAt'] = posts['createdAt'].apply(lambda x: int(x.timestamp()))

posts['date'] = posts['createdAt'] .map(lambda x: datetime.utcfromtimestamp(x).strftime('%Y-%m-%d'))
posts.set_index('date', inplace=True)
posts.head()

# Sort from latest to oldest
posts.sort_index(inplace=True,ascending=False)
posts.head()

# Check the number of unique users and reddit posts per subreddit/community 
users = posts[['communityName','username','body']].groupby(['communityName'],as_index=False).nunique()
users.sort_values(['body'],inplace=True, ascending=False)
users.reset_index(inplace=True, drop=True)
users

# Save the data to local disk
users.to_excel('UsersCount.xlsx')


# Check the daily distribution of number of posts published by all users in all the selected subreddits over time
distribution = posts[['body']].groupby(posts.index).nunique()
distribution.head()

distribution.tail()

# Save the time-series on local disk for further processing
distribution.to_csv('postsTS.csv')

# Initiate the CounteVectorizer and count the number of Keywords; mainly unigrams, bigrams and trigrams
# Rerun the cell by changing the ngram_range parameter for bigrams and trigrams once it is done running for unigrams

start = time.monotonic()

cv = CountVectorizer(ngram_range=(1,1),stop_words='english')  #Also use; ngram_range=(2,2) & (3,3) for bigrams & trigrams

cvFit = cv.fit_transform([str(posts['body'].values.tolist())])

wordList = cv.get_feature_names_out()
countList = cvFit.toarray().sum(axis=0)

dictionary = dict(zip(wordList,countList))
dictionary = dict(sorted(dictionary.items(), key = lambda x: x[1], reverse=True))

cvDf = pd.DataFrame(list(dictionary.items()), columns=['keywords','count'])

end = time.monotonic()

print(f"Total time taken: {timedelta(seconds=(end-start))}")

cvDf.head()

cvDf.tail()

# Finally save the "keywordsBigrams100.csv" and "keywordsTrigrams100.csv" files by repeating the same previous step
# Keywords with greater than or equal to 100 mentions only
cvDf[cvDf['count'] >= 100].to_csv("keywordsUnigrams100.csv")

# Count the number of unique Reddit users(ie authors)
posts['username'].nunique()
