==> Analysis/8-ClustersFullOrderPartialCorrelation.ipynb <==
%%time

# This notebook was used to perform full order partial correlation on the climate related themes

# Import required libraries
import pandas as pd
import numpy as np
import pingouin as pg
import os

new_working_directory = "/Users/fatimaq/Documents/Qualitative_reddit_analysis/Output"
os.chdir(new_working_directory)


# Load the time series data
df = pd.read_csv("Sample1ClustersProportion.csv")
df.set_index('date',drop=True,inplace=True)
df.head()

clusters = df.columns.to_list()
parCorr = pd.DataFrame()
for i in df.columns.to_list():
    for j in clusters:
        if i!=j:
            output = pg.partial_corr(data=df, x=f'{j}', y=f'{i}', covar=[k for k in clusters if k not in [i, j]]).round(3)
            output['-95%CI'] = output['CI95%'].map(lambda x: x[0])
            output['+95%CI'] = output['CI95%'].map(lambda x: x[1])
            output['y'] = i
            output['x'] = j            
            output.drop(['CI95%'],inplace=True,axis=1)
            parCorr = pd.concat([parCorr, output])

parCorr.head()    

parCorr.to_csv("ClustersParCorrOutput.csv",index=False)
